Update the logic around connectors so bridge-mode is only enabled by default for users that don't have SmartScale or other R/W Splitting features enabled. Do not push JGroups or TLS keystore files to connector-only servers CONT-1033
(cherry picked from commit f40910683873cc7493ba976b09a1d0317801bc35)

+++ builder/extra/tools/ruby-tpm/configure/deployment_steps/deployment_step_cluster.rb
@@ -414,7 +414,7 @@ EOF
           end
           
           connector_user_map = current_release_target_dir + '/tungsten-connector/conf/user.map'
-          if File.exists?(connector_user_map) && @config.getProperty(CONN_DELETE_USER_MAP) == "false"
+          if File.exists?(connector_user_map) && @config.getProperty(CONN_DELETE_USER_MAP) == "false" && @config.getProperty(ENABLE_CONNECTOR_BRIDGE_MODE) == "false"
             info("Copy the previous connector user map")
             FileUtils.cp(connector_user_map, prepare_dir + '/tungsten-connector/conf')
           end
